init project

+++ app/Http/Controllers/DanhMucController.php
@@ -3,7 +3,6 @@
 namespace App\Http\Controllers;
 
 use App\Models\DanhMuc;
-use App\Models\NhanVien;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 
@@ -22,15 +21,22 @@ class DanhMucController extends Controller
     {
         $login = Auth::guard('sanctum')->user();
         if ($login) {
-            DanhMuc::create([
-                "ten_danh_muc"          => $request->ten_danh_muc,
-                "slug_danh_muc"         => $request->slug_danh_muc,
-                "tinh_trang"            => 1,
-            ]);
-            return response()->json([
-                'status'        => true,
-                'message'       => 'Đã thêm mới danh mục thành công!'
-            ]);
+            if ($login->tinh_trang == 1) {
+                DanhMuc::create([
+                    "ten_danh_muc"          => $request->ten_danh_muc,
+                    "slug_danh_muc"         => $request->slug_danh_muc,
+                    "tinh_trang"            => 1,
+                ]);
+                return response()->json([
+                    'status'        => true,
+                    'message'       => 'Đã thêm mới danh mục thành công!'
+                ]);
+            } else {
+                return response()->json([
+                    'status'        => false,
+                    'message'       => 'Tài khoản của bạn đang tạm khóa!'
+                ]);
+            }
         }
     }
 
@@ -38,14 +44,21 @@ class DanhMucController extends Controller
     {
         $login = Auth::guard('sanctum')->user();
         if ($login) {
-            DanhMuc::where("id", $request->id)->update([
-                "ten_danh_muc"          => $request->ten_danh_muc,
-                "slug_danh_muc"         => $request->slug_danh_muc,
-            ]);
-            return response()->json([
-                'status'        => true,
-                'message'       => 'Đã cập nhật danh mục thành công!'
-            ]);
+            if ($login->tinh_trang == 1) {
+                DanhMuc::where("id", $request->id)->update([
+                    "ten_danh_muc"          => $request->ten_danh_muc,
+                    "slug_danh_muc"         => $request->slug_danh_muc,
+                ]);
+                return response()->json([
+                    'status'        => true,
+                    'message'       => 'Đã cập nhật danh mục thành công!'
+                ]);
+            } else {
+                return response()->json([
+                    'status'        => false,
+                    'message'       => 'Tài khoản của bạn đang tạm khóa!'
+                ]);
+            }
         }
     }
 
@@ -53,12 +66,19 @@ class DanhMucController extends Controller
     {
         $login = Auth::guard('sanctum')->user();
         if ($login) {
-            DanhMuc::where("id", $request->id)->delete();
+            if ($login->tinh_trang == 1) {
+                DanhMuc::where("id", $request->id)->delete();
 
-            return response()->json([
-                'status'        => true,
-                'message'       => 'Đã xóa danh mục thành công!'
-            ]);
+                return response()->json([
+                    'status'        => true,
+                    'message'       => 'Đã xóa danh mục thành công!'
+                ]);
+            } else {
+                return response()->json([
+                    'status'        => false,
+                    'message'       => 'Tài khoản của bạn đang tạm khóa!'
+                ]);
+            }
         }
     }
 
@@ -66,17 +86,24 @@ class DanhMucController extends Controller
     {
         $login = Auth::guard('sanctum')->user();
         if ($login) {
-            $danh_muc = DanhMuc::where("id", $request->id)->first();
-            if ($danh_muc) {
-                $danh_muc->tinh_trang = !$danh_muc->tinh_trang;
-                $danh_muc->save();
-                return response()->json([
-                    'status'        => true,
-                ]);
+            if ($login->tinh_trang == 1) {
+                $danh_muc = DanhMuc::where("id", $request->id)->first();
+                if ($danh_muc) {
+                    $danh_muc->tinh_trang = !$danh_muc->tinh_trang;
+                    $danh_muc->save();
+                    return response()->json([
+                        'status'        => true,
+                    ]);
+                } else {
+                    return response()->json([
+                        'status'        => false,
+                        'message'       => 'Danh mục không tồn tại!'
+                    ]);
+                }
             } else {
                 return response()->json([
                     'status'        => false,
-                    'message'       => 'Danh mục không tồn tại!'
+                    'message'       => 'Tài khoản của bạn đang tạm khóa!'
                 ]);
             }
         }
